Clean the home page, add the irregular-verbs page
This commit cleans the home page and add a button to go to the new view
called IrregularVerbs in dutch, this vue is a form view in which the
user will must fill in the inputs based on prefill input. For example,
we have an irregular verb and we prefill the input for the
translation in french. Then, the user will fill the infinitive input,
past simple input,  past simple in plurial form input and the
participle past input. If he correctly fill the inputs then we continue
to the next irregular verb, etc.