fix: errors when checkAnswers, counter good or bad answers
When the verb can has more than one possibilities of translation, if the
user wants to give all possibilities the system doesn't check in any
order because it waited for only one solution. Moreover, when the user
has made more than 10 exercices, the number of bad answers disappears
and only the number of good answers are shown.

This commit changes the checkAnswers method to check all possibilities
that the user gives. And it fix the error to display the number of good
and bad answers.