fix the bad answer give coin rather than close icon
Before this commit, when the user made a bad answer, the system gives
him a coin rather than close icon. It is because the answer is no longer
a boolean value but an object containing correct property to know if the
answer is correct or not.

This commit uses the correct property in the answer object to know if
the answer is correct or not and then gives a coin or a close icon.