Fix undeterministic bug when select random verb + add style resume page
Before this commit, when the user does the exercices, sometimes he gets
the same verbs and the next of exercices bugs since it should not select a
same verb in same exercice.

This commit reviews the code to fix the error.
Moreover, this commit adds some styles on the resume page. To display
red line for the bad answers and green line for the expected verbs.