Use the setup sugar for all vue components
Before this commit we need to use the defineComponent function to create
the vue component or use the setup in the script tag but we need to
return all the variables and functions to use them in the component
template.

This commit uses the new update of Vue js to use the setup in script.
Now all the functions and variables defined in the script are
automatically returned by vue. No need to explicitly write return
object.