Refatoring the answer given in IrregularVerb view and resume error feature
A new hook is created to generalize the answer given into an exercise.
Here just the Irregular verbs exercise is implemented but we will add
an exercise for the vocabulary. To not duplicate the code in both view,
this hook will be used in both views.

The resume error has been improved to manage the vocabulary exercise.

Since the hook stored the answers, the state in vuex has been updated to
no longer store the answers.