Refactoring the vuex store
Before this commit, we don't have the autocompletion for dispatch and
commit. Because the type are not created and use to create the real
store type.

This commit creates the store type and return the store with this type.
Then, we have the autocompletion when we use the store in any components
and views.