Split Header and row for vocabulary and irregular verb in components
Before this commit, the header and the rows of vocabulary and irregular
verb are explicit write and components for ResumeError view.
However, the template for header can be done in a component for
vocabulary and other one for irregular verb to use these compoenents in
others views. Same idea for the row.

This commit adds Vocabulary{Header,Row} components for vocabulary and
IrregularVerbHeader for irregular verbs.