Define vuex module for irregular verbs feature
Before this commit, the irregular verbs store is defined in the main
vuex store. But this store should be in its own module as vocabulary
module to have a module for a feature.

The main store will store the language selected by the user. This
language will be used in the vocabulary and irregular verb modules to
call the api in the correct route.

Moreover, the Language interface definition is moved in models file.