Fix some issues to get resume errors and avoid select language field
Before this commit, the past simple in plural are not correctly selected
during the irregular verbs exercises because the name of the field is no
longer the same in the API. It is `past_simple2` instead of
`past_simple_2`. Moreover, since we add the language to manage many
languages (english and dutch). The field randomly selected to start the
exercise is the one containing the id of the language of the vocabulary
or the word.

This commit renames the field of the past simple in plural to fit with
the one in the API and it ignores the `language_id` field during the
selection of a field to correctly begin the exercise.