Review the MainContainer component
Before this commit, we create some layout to have the same appearance
than MainContainer without menu content and the center style.

This commit adds 2 options in the MainContainer to allow to reuse this
component instead of creating a new layout. This is the MainLayout
component will be deprecated after this commit.