Set language into the vue store
Before this commit, we use the useLang hook to get the language based on
the params lang contained in the vue router.

This commit deprecated this hook and set the lang when we enter to the
route `/nl` or `/en` to store and use the store to get the correct api
route based on the language chosen by the user.
Moreover, it uses the getter getLang to get the language as the hook
does.