fetch vocabularies based on the category chosen
Before this commit, when the user wants to launch an vocabulary
exercise for a specific category, he gets the exercise with all the
vocabularies even those contained in others categories.

This commit changes the fetchData function to add parameter for the
action to call. This new parameter will contain the id of the category
when we need to fetch the vocabularies for a specific category.