fix wrong selected field name of vocabulary exercise.
Before this commit, when the user goes to the vocabulary exercises, he
can have only blank inputs in an exercise. So, he isn't able to do the
exercise since he cannot know which word he must find.

This commit changes the useExercise hook to correctly select a field
name to show either the word in the language selected or the
translation. Then, the user will always be able to know which word he
must to translate or find.