Add the edition for irregular verbs
Before this commit, it was impossible to the user to add some irregular
verbs. So, in english no irregular verb is added and it was impossible
to do an exercise since there is no data.

This commit allows the user to see the list of irregular verbs as the
list of vocabularies (but without category since the irregular verbs
don't need to be classified into any categories). It also allows to add
or edit an irregular verb.